\name{bayesMargEffF}
\alias{bayesMargEffF}
\alias{bayesMargEffF.stanreg}
\alias{bayesMargEffF.data.frame}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bayesian marginal effects on the mean scale.
}
\description{
Computes Bayesian marginal effects on the mean scale for models fit using the package "rstanarm". Marginal effects can be averaged over the values of the covariates in the data (average marginal effects), or the covariates can be held at their means (marginal effects at the means). Also, "at" values can be specified to fix covariates at particular values. The marginal effects can be specified in terms of discrete changes or instantaneous rates of change.
}
\usage{

bayesMargEffF(x, ...)

\method{bayesMargEffF}{stanreg}(x,
              marginal_effect,
              start_value,
              end_value,
              n_draws         = 2000,
              ci              = .95,
              hdi_interval    = TRUE,
              centrality      = 'mean',
              digits          = 4,
              at              = NULL,
              at_means        = FALSE,
              h               = .0001,
              ...)

\method{bayesMargEffF}{data.frame}(x,
              model_data,
              model_formula,
              link_function,
              marginal_effect,
              start_value,
              end_value,
              model_offset    = NULL,
              n_draws         = 2000,
              ci              = .95,
              hdi_interval    = TRUE,
              centrality      = 'mean',
              digits          = 4,
              at              = NULL,
              at_means        = FALSE,
              h               = .0001,
              ...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{x}{
A model object of class "stanreg" or an object of class "data.frame" that contains the joint posterior.
}
\item{model_data}{
Object of class "data.frame" that contains the data used to fit the model. 
}
\item{model_formula}{
Object of class "formula". The formula must not contain an offset; instead, include any offsets in the "model_offset" argument.
}
\item{link_function}{
The link function used to fit the model. Current options are "logit", "probit", "cloglog", "log", "inverse", "sqrt", or "identity".
}
\item{marginal_effect}{
A character vector of terms that you want to estimate the marginal effect for.
}
\item{start_value}{
The start value(s) for the marginal effect(s). Must be a list when specifying multiple values. Specify "instantaneous" when you want to calculate an approximate instantaneous rate of change.
}

\item{end_value}{
The end value(s) for the marginal effect(s). Must be a list when specifying multiple values. Specify "instantaneous" when you want to calculate an approximate instantaneous rate of change.
}
\item{model_offset}{
Offset used in the model. Must be a scalar value.
}
\item{n_draws}{
The number of draws to take from the joint posterior distribution.
}
\item{ci}{
The level for the credible intervals.
}
\item{hdi_interval}{
If TRUE, the default, computes the highest density credible interval. If FALSE, computes the equal-tailed interval.
}
\item{centrality}{
Centrality measure for the posterior distribution. Options are "mean" or "median".
}
\item{digits}{
The number of digits to report in the summary table.
}
\item{at}{
Optional list of covariate values to estimate the marginal effect at.
}
\item{at_means}{
If FALSE, the default, the marginal effects are averaged across the rows of the data. If TRUE, the covariate values that are not specified in the "at" argument are held at their means.
}
\item{h}{
Step size for calculating the approximate instantaneous rate of change.
}
\item{...}{
Arguments passed to method.
}
}
\value{
A list of class "bayesmeanscale_marg" with the following components: 

\item{diffTable}{
summary table of the marginal effects
}
\item{diffDraws}{
posterior draws of the marginal effects
}

}
\details{
The following families of fixed-effect models fit using "rstanarm" are supported: 'beta', 'binomial', 'Gamma', 'gaussian', 'neg_binomial_2', and 'poisson.'
}
\author{
David Dalenberg
}
\references{
Agresti, Alan. 2013. \emph{Categorical Data Analysis}. Third Edition. New York: Wiley

Long, J. Scott and Sarah A. Mustillo. 2018. "Using Predictions and Marginal Effects to Compare Groups in Regression Models for Binary Outcomes." \emph{Sociological Methods & Research} 50(3): 1284-1320.

Mize, Trenton D. 2019. "Best Practices for Estimating, Interpreting, and Presenting Non-linear Interaction Effects." \emph{Sociological Science} 6: 81-117. 
}
\examples{

\donttest{

## Logit model ##

if(require(rstanarm)){

m1 <- stan_glm(switch ~ dist + educ + arsenic + assoc, 
               data    = rstanarm::wells, 
               family  = binomial, 
               refresh = 0, 
               iter    = 500)

# marginal effect averaging over sample values of covariates #

bayesMargEffF(m1, 
              marginal_effect = 'arsenic', 
              start_value     = 2.2, 
              end_value       = .82,
              n_draws         = 500)
              
}
              
}

}
